pyproject.toml
[project]
name = "cs778"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "einops>=0.8.1",
    "google-generativeai>=0.8.5",
    "peft>=0.18.0",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "trl>=0.25.1",
]
